requirements.txt
opentelemetry-distro
yfinance
fastapi
